refactor!: move pkg/dag to internal/dag
Keep Container.Graph() hackable via a ServiceGraph alias and re-exported DAG errors from pal, without shipping a public DAG utility package.
